% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manu.R
\name{get_pal}
\alias{get_pal}
\title{Select Manu palette}
\usage{
get_pal(name)
}
\arguments{
\item{name}{Name of desired palette. Choices are: \code{Hihi}, \code{Hoiho}, \code{Kaka}, \code{Kakariki}, \code{Kea}, \code{Kereru}, \code{Kereru_orig}, \code{Korimako}, \code{Korora}, \code{Kotare}, \code{Putangitangi}, \code{Takahe}, \code{Takapu}, \code{Titipounamu}, \code{Tui}, \code{Pepetuna} and \code{Pohutukawa}}
}
\value{
A vector of colours.
}
\description{
This function returns colour palettes derived from photos of birds native to New Zealand
}
\examples{
get_pal("Hoiho")
get_pal("Takahe")
get_pal("Kereru")
}
