% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manu.R
\docType{data}
\name{manu_palettes}
\alias{manu_palettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 17.
}
\usage{
manu_palettes
}
\description{
These are colour palettes derived from photos of birds native to New Zealand. Use \code{\link{get_pal}} extract palettes.
}
\keyword{datasets}
